adds assert and float type checks
This commit introduces a negate function which simplifies the implementations of things like cpp_is_not_float. It refactors cpp_is_not_empty and cpp_is_not_int in terms of negate. It also introduces the assert function, which wraps the process of crashing the program if a condition is not met while CMakePP is running in debug mode. Finally, this commit adds type checks for floats.