Added ref labels to 'classes'and 'types' feature docs
These ref labels make it much easier to link the user to further info about
the features being used, which is especially useful in the first few examples
present in "Getting Started". I tried to reference them in a way that is
unique for the entire project by roughly doing
'_<path-to-file-from-source-dir>-<filename>-<approximate-header>`.